AppScreens + SeriesScreen

+++ app/data/src/commonMain/kotlin/ru/slartus/moca/data/api/tmdb/models/Tv.kt
@@ -13,8 +13,8 @@ data class Tv(
     val original_language: String?,
     val original_name: String?,
     val overview: String?,
-    val popularity: Double?,
+    val popularity: Float?,
     val poster_path: String?,
-    val vote_average: Double?,
+    val vote_average: Float?,
     val vote_count: Int?
 )
